major changes

+++ blogapp/views.py
@@ -1,6 +1,7 @@
-from django.shortcuts import render, redirect
-from django.contrib.auth.decorators import login_required
+from django.shortcuts import render, redirect, get_object_or_404
+from django.contrib.auth.decorators import login_required, permission_required
 from django.conf import settings
+from users.decorators import allowed_users
 
 import os
 import shutil
@@ -9,10 +10,35 @@ import shutil
 from .models import Topic, Image
 from .forms import CommentForm, TopicForm, ImageForm
 
+def home(request):
+    return redirect ('/0/')
 
-def index(request):
-    topics = Topic.objects.order_by('-date_created')
-    context = {'topics': topics}
+
+def index(request, page_nr):
+    
+    all_topics = Topic.objects.order_by('-date_created')
+    numbers_of_topic_for_single_page = 3
+    tpp = numbers_of_topic_for_single_page
+     
+    if len(all_topics) - (page_nr * tpp + tpp) >= 0:
+        visible_topics = all_topics[page_nr*tpp : page_nr*tpp+tpp]
+    elif len(all_topics) - (page_nr * tpp + tpp) < tpp > 0:
+        visible_topics = all_topics[page_nr*tpp : ]
+
+    class Page():
+        def __init__(self) -> None:
+            self.nr = page_nr
+            self.last_nr = int(len(all_topics) / (tpp))
+            self.next_nr = self.nr + 1
+            if self.nr >0:
+                self.prev_nr = self.nr - 1
+            else:
+                self.prev_nr = 0
+    
+    
+    page = Page()
+
+    context = {'topics': visible_topics, 'page': page,}
     return render(request, 'blogapp/index.html', context)
 
 
@@ -50,7 +76,8 @@ def gallery(request):
 
 
 def topic(request, topic_id):
-    topic = Topic.objects.get(id = topic_id)
+    # topic = Topic.objects.get(id = topic_id)
+    topic =    get_object_or_404(Topic, id=topic_id)
     comments = topic.comment_set.order_by('-date_created')
     images = topic.image_set.all()
 
@@ -66,27 +93,25 @@ def topic(request, topic_id):
             new_comment.save()
         return redirect('blogapp:topic', topic.id)
    
-    context = {
-        'topic': topic, 'comments': comments, 'form': form,'images':images,
-         }
+    context ={'topic': topic, 'comments': comments, 'form': form,'images':images,}
     return render(request, 'blogapp/topic.html', context)
 
 
 @login_required
+@permission_required('blogapp.add_topic', raise_exception=True)
+# @allowed_users(allowed = ['admins', 'moderators'])
 def new_topic(request):
     if request.method != 'POST':
         form = TopicForm()
     else:
-        form = TopicForm(data=request.POST)
+        form = TopicForm(request.POST, request.FILES)
         if form.is_valid():
             new_topic = form.save(commit=False)
             new_topic.owner= request.user
             new_topic.save()
             if not new_topic.title in os.listdir(settings.MEDIA_ROOT):
                 os.mkdir(os.path.join(settings.MEDIA_ROOT, new_topic.title))
-            
-                pass
-            return redirect('blogapp:index')
+            return redirect('blogapp:home')
 
     context = {'form': form}
 
@@ -95,26 +120,29 @@ def new_topic(request):
 
 @login_required
 def edit_topic(request, topic_id):
-    topic = Topic.objects.get(id=topic_id)
+    topic = get_object_or_404(Topic, id=topic_id)
 
     if request.method != 'POST':
         form= TopicForm(instance=topic)
     else:
-        form = TopicForm(instance=topic, data=request.POST)
+        form = TopicForm(request.POST, request.FILES, instance=topic)
         if form.is_valid:
             form.save()
             return redirect('blogapp:topic', topic_id=topic.id)
+
     context = {'topic':topic, 'form':form}
+
     return render(request, 'blogapp/edit_topic.html', context)
 
 
 @login_required
 def delete_topic(request, topic_id):
-    topic = Topic.objects.get(id=topic_id)
+    topic = get_object_or_404(Topic, id=topic_id)
     if request.method =='POST':
         topic.delete()
-        shutil.rmtree(os.path.join(settings.MEDIA_ROOT, topic.title))
-        return redirect('blogapp:index')
+        if  topic.title in os.listdir(settings.MEDIA_ROOT):                
+            shutil.rmtree(os.path.join(settings.MEDIA_ROOT, topic.title))
+        return redirect('blogapp:home')
 
     context = {'topic':topic}
     
@@ -123,14 +151,17 @@ def delete_topic(request, topic_id):
 
 @login_required
 def add_image(request, topic_id):
-    topic = Topic.objects.get(id=topic_id)
+    topic = get_object_or_404(Topic, id=topic_id)
 
     if request.method != 'POST':
         form= ImageForm()
     else:
         form = ImageForm(request.POST, request.FILES)
-        if form.is_valid:
+        if form.is_valid:             
+            checkbox_value = request.POST.get('show_in_gallery')
             new_image = form.save(commit=False)
+            if not checkbox_value:
+                new_image.show_in_gallery = False
             new_image.topic= topic
             new_image.save()
             return redirect('blogapp:topic', topic_id=topic.id)
